Support for handing JSON protobufs
Consumers now check the message metadata for a message format type which is
used to decode the message. If one cannot be found, it falls back to assuming
the message is a Base64 encoded binary protobuf. This allows us to support
JSON protobufs without breaking existing message contracts.

This is a bit ugly due to both the way we mock things and the need to
pass the message format down through several layers of functions.
Planning on getting rid of the unnecessary mocks and redundant layers in
a separate PR.