JSON protobuf encoding
Support for publishing messages in JSON protobuf format. The
`#publish` macro now takes an optional second argument which
indicates the message format to use. The value can be either
`"json_protobuf"` or `"binary_protobuf"`. If the argument is
omitted, `"binary_protobuf"` is used.

The message format argument is then passed down through the
call chain to the `Payload` module.